change tasks_count to size_t *

+++ server/libs/packet.h
@@ -33,7 +33,7 @@ extern pthread_mutex_t task_lock;
 extern pthread_cond_t tasks_cond[MAXTASKSCOUNT];
 extern pthread_cond_t add_task_cond;
 extern char **tasks_name;
-extern int *tasks_count;
+extern size_t *tasks_count;
 
 
 typedef struct list{
